safari-600.7-branch/LayoutTests/platform/mac/accessibility/title-attribute-used-as-last-resort-expected.txt

text text
This tests that the title attribute is used as the last resort for description when there is no other option available, but otherwise used as help text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.title is 'AXTitle: text'
PASS accessibilityController.focusedElement.helpText is 'AXHelp: TITLE'
PASS accessibilityController.focusedElement.title is 'AXTitle: text'
PASS accessibilityController.focusedElement.helpText is 'AXHelp: HELP'
PASS accessibilityController.focusedElement.description is 'AXDescription: TITLE'
PASS accessibilityController.focusedElement.helpText is 'AXHelp: '
PASS accessibilityController.focusedElement.description is 'AXDescription: TITLE'
PASS accessibilityController.focusedElement.helpText is 'AXHelp: HELP'
PASS accessibilityController.focusedElement.description is 'AXDescription: LABEL'
PASS accessibilityController.focusedElement.helpText is 'AXHelp: HELP'
PASS successfullyParsed is true

TEST COMPLETE
